Curate api models restructure
