+++ spb_curate/abstract/__init__.py
@@ -1,2 +1,3 @@
 from spb_curate.abstract import api, superb_ai_object
 from spb_curate.abstract.api import *  # noqa
+from spb_curate.abstract.superb_ai_object import SuperbAIObject  # noqa
