+++ spb_curate/abstract/api/__init__.py
@@ -1,5 +1,5 @@
 from spb_curate.abstract.api import resource
-from spb_curate.abstract.api.resource import (
+from spb_curate.abstract.api.resource import (  # noqa
     CreateResource,
     DeleteResource,
     PaginateResource,
