+++ spb_curate/curate/api/settings.py
@@ -1,0 +1,2 @@
+FETCH_PAGE_LIMIT = 100
+UPLOAD_IMAGE_FILE_BYTES_MAX = 20000000  # 20MB
